move exception checking under es.py
